+++ src/app/student/log/page.tsx
@@ -1,0 +1,7 @@
+import { StudentLogPage } from '@/page/manage';
+
+const Page = () => {
+  return <StudentLogPage />;
+};
+
+export default Page;
