+++ src/app/trainer/manage/[memberId]/log/[LogId]/page.tsx
@@ -1,11 +1,11 @@
-import { StudentLogDetailPage } from '@/page/manage/ui/StudentLogDetailPage';
+import { TrainerLogDetailPage } from '@/page/manage';
 
 interface Props {
   params: { logId: number; memberId: number };
 }
 
 const Page = ({ params }: Props) => {
-  return <StudentLogDetailPage logId={params.logId} memberId={params.memberId} />;
+  return <TrainerLogDetailPage logId={params.logId} memberId={params.memberId} />;
 };
 
 export default Page;
