+++ src/app/trainer/manage/[memberId]/log/page.tsx
@@ -1,11 +1,11 @@
-import { StudentLogPage } from '@/page/manage';
+import { TrainerLogPage } from '@/page/manage';
 
 interface Props {
   params: { memberId: number };
 }
 
 const Page = ({ params }: Readonly<Props>) => {
-  return <StudentLogPage memberId={params.memberId} />;
+  return <TrainerLogPage memberId={params.memberId} />;
 };
 
 export default Page;
