+++ src/page/manage/index.ts
@@ -1,10 +1,13 @@
 export { NotRegisteredStudentListPage } from './ui/NotRegisteredStudentListPage';
 export { StudentCourseDetailPage } from './ui/StudentCourseDetailPage';
-export { StudentDetailPage } from './ui/StudentDetailPage';
 export { StudentEditMemo } from './ui/StudentEditMemo';
 export { StudentListPage } from './ui/StudentListPage';
+export { StudentLogDetailPage } from './ui/StudentLogDetailPage';
 export { StudentLogPage } from './ui/StudentLogPage';
 export { TrainerAppendStudentPage } from './ui/TrainerAppendStudentPage';
 export { TrainerCreateLogPage } from './ui/TrainerCreateLogPage';
 export { TrainerFeedbackPage } from './ui/TrainerFeedbackPage';
 export { TrainerInvitePage } from './ui/TrainerInvitePage';
+export { TrainerLogDetailPage } from './ui/TrainerLogDetailPage';
+export { TrainerLogPage } from './ui/TrainerLogPage';
+export { TrainerStudentDetailPage } from './ui/TrainerStudentDetailPage';
