+++ src/page/manage/ui/StudentLogDetailPage.tsx
@@ -1,0 +1,18 @@
+import { Layout } from '@/widget';
+
+interface Props {
+  logId: number;
+  memberId: number;
+}
+
+const StudentLogDetailPage = ({ memberId, logId }: Props) => {
+  console.log(memberId, logId);
+  return (
+    <Layout>
+      <Layout.Header></Layout.Header>
+      <Layout.Contents></Layout.Contents>
+    </Layout>
+  );
+};
+
+export { StudentLogDetailPage };
