+++ src/page/manage/ui/TrainerLogDetailPage/Header.tsx
